backend/requirements.txt
selenium
python-dotenv
pandas
bs4
webdriver_manager
setuptools

typing-extensions
Flask-SQLAlchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
playwright
gunicorn
gitpython
docx2pdf
PyPDF2>=3.0
psycopg2-binary
# psycopg2>=2.9
pillow 
pytesseract
flask
flask-cors
werkzeug
wheel
image


